refactor imported class in user repository test, email controller test, and user service test 👌👌👌
