+++ src/test/java/com/security/jdbc/service/UserServiceTest.java
@@ -1,15 +1,14 @@
 package com.security.jdbc.service;
 
 
-import com.security.jdbc.repository.UserRepository;
+import com.security.jdbc.authentication.service.UserService;
+import com.security.jdbc.authentication.repository.UserRepository;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.password.PasswordEncoder;
